requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.3
django-background-tasks==1.2.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-compat==1.0.15
django-constance==2.9.0
django-cors-headers==4.3.1
django-environ==0.8.1
django-filter==24.1
django-timezone-field==6.1.0
djangorestframework==3.15.0
environ==1.0
google-api-core==2.18.0
google-apps-meet==0.1.6
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
kombu==5.3.6
Markdown==3.6
oauthlib==3.2.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stripe==8.8.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
